requirements.txt
networkx==3.2.1
matplotlib==3.8.2
openai==1.55.3
tqdm==4.66.1
python-dotenv==1.0.0
